# docs/swagger.yaml
basePath: /api/v1
definitions:
  authentications.AuthIn:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  authentications.AuthOut:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      token:
        $ref: '#/definitions/authentications.Token'
    type: object
  authentications.InitUserPasswordIn:
    properties:
      password:
        type: string
      user_id:
        type: string
    required:
    - password
    - user_id
    type: object
...
  authentications.ResetPasswordIn:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  authentications.Token:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
    type: object
...
  utils.HttpResponse-any:
    properties:
      code_error:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.HttpResponse-authentications_AuthOut:
    properties:
      code_error:
        type: integer
      data:
        $ref: '#/definitions/authentications.AuthOut'
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Service d'authentification.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CMagic Auth
  version: 1.0.0
paths:
...
  /auth/init_password:
    put:
      description: Initialisation du mot de passe utilisateur
      parameters:
      - description: Body data
        in: body
        name: init_password
        required: true
        schema:
          $ref: '#/definitions/authentications.InitUserPasswordIn'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/utils.HttpResponse-any'
      summary: Initialisation du mot de passe utilisateur
      tags:
      - Authentications
  /auth/login:
    post:
      description: Authentifier un utlisateur
      parameters:
      - description: Détails de l'utilisateur
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/authentications.AuthIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.HttpResponse-authentications_AuthOut'
      summary: Authentifier un utlisateur
      tags:
      - Authentications
...
  /auth/reset_password:
    put:
      description: Changer de mot de passe
      parameters:
      - description: Détails de l'utilisateur
        in: body
        name: reset_password
        required: true
        schema:
          $ref: '#/definitions/authentications.ResetPasswordIn'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/utils.HttpResponse-authentications_AuthOut'
      summary: Changer de mot de passe
      tags:
      - Authentications
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"